backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==23.0.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Caching & Sessions
redis==5.1.1
hiredis==3.0.0
fastapi-cache2[redis]==0.2.2

# Authentication & Security
cryptography==43.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# HTTP Client (Async)
httpx==0.27.2
aiohttp==3.10.10

# Task Queue & Scheduling
celery==5.4.0
apscheduler==3.10.4

# WebSocket
python-socketio==5.11.4

# AI & NLP
openai==1.51.2
langchain==0.3.7
langchain-openai==0.2.5
tiktoken==0.8.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.4

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.8
pytest-cov==5.0.0
faker==30.3.0

# Code Quality
ruff==0.7.0
mypy==1.13.0

# Monitoring & Logging
python-json-logger==2.0.7

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# File Storage
boto3==1.35.36
minio==7.2.9

# Data Validation
email-validator==2.2.0
